Reapply "Avoid or make integer precision conversion explicit"
This reverts commit 6c43eb76d1a4da50ce7fdb28f2901d7d75d1f9fd.

+++ src/int_utils.h
@@ -210,7 +210,7 @@ public:
     static constexpr inline int TopBits(I val) {
         static_assert(Count > 0, "BitsInt::TopBits needs Count > 0");
         static_assert(Count <= BITS, "BitsInt::TopBits needs Offset <= BITS");
-        return val >> (BITS - Count);
+        return static_cast<int>(val >> (BITS - Count));
     }
 
     static inline constexpr I CondXorWith(I val, bool cond, I v) {
